+++ module/helpers/attachmentHelper.mjs
@@ -136,3 +136,81 @@ export async function handleAttachmentEffectsOnEquipChange({ parentItem, newEqui
         }
     }
 }
+
+/**
+ * Prepare attachment context data for rendering
+ * @param {Item} parentItem - The item (armor/weapon) that has attachments
+ * @returns {Promise<Object[]>} Array of attachment data objects
+ */
+export async function prepareAttachmentContext(parentItem) {
+    const attachedUUIDs = parentItem.system.attached;
+    return await Promise.all(
+        attachedUUIDs.map(async uuid => {
+            const item = await fromUuid(uuid);
+            return {
+                uuid: uuid,
+                name: item?.name || 'Unknown Item',
+                img: item?.img || 'icons/svg/item-bag.svg'
+            };
+        })
+    );
+}
+
+/**
+ * Add an attachment to an item via drag and drop
+ * @param {Object} options - Configuration options
+ * @param {Item} options.parentItem - The item (armor/weapon) that the item is being attached to
+ * @param {Item} options.droppedItem - The item being attached
+ * @param {string} options.parentType - Type of parent item ("armor" or "weapon")
+ * @returns {Promise<void>}
+ */
+export async function addAttachmentToItem({ parentItem, droppedItem, parentType }) {
+    const currentAttached = parentItem.system.attached;
+    const newUUID = droppedItem.uuid;
+    
+    if (currentAttached.includes(newUUID)) {
+        ui.notifications.warn(`${droppedItem.name} is already attached to this ${parentType}.`);
+        return;
+    }
+    
+    const updatedAttached = [...currentAttached, newUUID];
+    
+    await parentItem.update({
+        'system.attached': updatedAttached
+    });
+    
+    // Copy effects from attached item to actor (only if parent item is equipped)
+    const actor = parentItem.parent;
+    if (actor && droppedItem.effects.size > 0 && parentItem.system.equipped) {
+        await copyAttachmentEffectsToActor({
+            parentItem,
+            attachedItem: droppedItem,
+            attachedUuid: newUUID,
+            parentType
+        });
+    }
+}
+
+/**
+ * Remove an attachment from an item and clean up its effects
+ * @param {Object} options - Configuration options
+ * @param {Item} options.parentItem - The item (armor/weapon) that the item is attached to
+ * @param {string} options.attachedUuid - UUID of the attached item being removed
+ * @param {string} options.parentType - Type of parent item ("armor" or "weapon")
+ * @returns {Promise<void>}
+ */
+export async function removeAttachmentFromItem({ parentItem, attachedUuid, parentType }) {
+    const currentAttached = parentItem.system.attached;
+    
+    // Remove the attachment from the parent item's attached array
+    await parentItem.update({
+        'system.attached': currentAttached.filter(uuid => uuid !== attachedUuid)
+    });
+    
+    // Remove any effects that came from this attachment
+    await removeAttachmentEffectsFromActor({
+        parentItem,
+        attachedUuid,
+        parentType
+    });
+}
